# openapi.yml
openapi: 3.1.0
info:
  title: Gugu Admin API
  version: 0.1.0
  description: |
    SKU 적재, 가격 갱신, AliExpress 토큰 관리를 위한 어드민 배치 API 명세
servers:
  - url: https://lees-admin.duckdns.org
    description: 운영 서버
  - url: http://localhost:8700
    description: 로컬 개발 서버
...
paths:
...
  /v1/batch/load-hot-products:
    post:
      tags:
        - batch
      security:
        - AdminApiKey: []
      summary: 인기 상품 적재
      description: |
        AliExpress 인기 상품을 KRW, USD 각각 조회하여 product_variant, product_price_history, product_price_snapshot에 적재합니다.
        대표 통화(KRW) 기준으로만 product 테이블에 신규 상품을 생성합니다.
        이미 존재하는 상품은 product insert를 건너뜁니다. SKU는 별도 enrich-skus 엔드포인트로 적재합니다.
      operationId: loadHotProducts
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HotProductLoadRequest'
      responses:
        '202':
          description: 인기 상품 적재 작업 시작됨
          headers:
            Cache-Control:
              $ref: '#/components/headers/Cache-Control'
            Pragma:
              $ref: '#/components/headers/Pragma'
            Expires:
              $ref: '#/components/headers/Expires'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AcceptedMessageResponse'
        '400':
          description: 잘못된 요청
          headers:
            Cache-Control:
              $ref: '#/components/headers/Cache-Control'
            Pragma:
              $ref: '#/components/headers/Pragma'
            Expires:
              $ref: '#/components/headers/Expires'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: 관리자 인증 헤더 누락
          headers:
            Cache-Control:
              $ref: '#/components/headers/Cache-Control'
            Pragma:
              $ref: '#/components/headers/Pragma'
            Expires:
              $ref: '#/components/headers/Expires'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/batch/enrich-skus/hot-products:
    post:
      tags:
        - batch
      security:
        - AdminApiKey: []
      summary: 인기 상품 SKU 보강
      description: SKU가 없는 인기 상품에 대해 Dropshipping API로 SKU를 적재합니다.
      operationId: enrichHotProducts
      responses:
        '202':
          description: SKU 보강 작업 시작됨
          headers:
            Cache-Control:
              $ref: '#/components/headers/Cache-Control'
            Pragma:
              $ref: '#/components/headers/Pragma'
            Expires:
              $ref: '#/components/headers/Expires'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AcceptedMessageResponse'
        '401':
          description: 관리자 인증 헤더 누락
          headers:
            Cache-Control:
              $ref: '#/components/headers/Cache-Control'
            Pragma:
              $ref: '#/components/headers/Pragma'
            Expires:
              $ref: '#/components/headers/Expires'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/batch/enrich-skus/all:
    post:
      tags:
        - batch
      security:
        - AdminApiKey: []
      summary: 전체 상품 SKU 보강
      description: 모든 상품에 대해 Dropshipping API로 SKU를 적재합니다.
      operationId: enrichAllProducts
      responses:
        '202':
          description: SKU 보강 작업 시작됨
          headers:
            Cache-Control:
              $ref: '#/components/headers/Cache-Control'
            Pragma:
              $ref: '#/components/headers/Pragma'
            Expires:
              $ref: '#/components/headers/Expires'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AcceptedMessageResponse'
        '401':
          description: 관리자 인증 헤더 누락
          headers:
            Cache-Control:
              $ref: '#/components/headers/Cache-Control'
            Pragma:
              $ref: '#/components/headers/Pragma'
            Expires:
              $ref: '#/components/headers/Expires'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/batch/update-product-prices:
    post:
      tags:
        - batch
      security:
        - AdminApiKey: []
      summary: 가격 갱신 배치
      description: |
        대상 상품들의 가격을 AliExpress API에서 다시 조회하여 갱신합니다.
        일반 상품은 product의 대표 통화 기준 1개 locale만 갱신합니다.
        `collection_source = HOT_PRODUCT_QUERY` 상품만 KRW/KO, USD/EN 각각 조회하여 product_variant와 product_price_history, product_price_snapshot을 갱신합니다.
        product 마스터 테이블은 가격을 저장하지 않고 title, image, product_url, last_collected_at 같은 메타만 유지합니다.
      operationId: updateProductPrices
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PriceUpdateRequest'
            examples:
              allProducts:
                summary: 전체 상품 대상
                value: {}
              hotProductsOnly:
                summary: 인기 상품만
                value:
                  collection_source: HOT_PRODUCT_QUERY
                  currencies:
                    - KRW
                    - USD
                  requested_by: admin
              filteredByMarket:
                summary: 특정 시각 이전 수집된 AliExpress 상품
                value:
                  market: ALIEXPRESS
                  collected_before: '2026-04-01T00:00:00Z'
      responses:
        '202':
          description: 가격 갱신 배치 시작됨
          headers:
            Cache-Control:
              $ref: '#/components/headers/Cache-Control'
            Pragma:
              $ref: '#/components/headers/Pragma'
            Expires:
              $ref: '#/components/headers/Expires'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PriceUpdateAcceptedResponse'
        '400':
          description: 잘못된 요청
          headers:
            Cache-Control:
              $ref: '#/components/headers/Cache-Control'
            Pragma:
              $ref: '#/components/headers/Pragma'
            Expires:
              $ref: '#/components/headers/Expires'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: 관리자 인증 헤더 누락
          headers:
            Cache-Control:
              $ref: '#/components/headers/Cache-Control'
            Pragma:
              $ref: '#/components/headers/Pragma'
            Expires:
              $ref: '#/components/headers/Expires'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: 프리뷰 실패
          headers:
            Cache-Control:
              $ref: '#/components/headers/Cache-Control'
            Pragma:
              $ref: '#/components/headers/Pragma'
            Expires:
              $ref: '#/components/headers/Expires'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/batch/update-sku-snapshots:
    post:
      tags:
        - batch
      security:
        - AdminApiKey: []
      summary: SKU snapshot 갱신 배치
      description: |
        대상 상품들의 SKU 가격을 AliExpress Dropshipping API에서 다시 조회하여 `sku_price_snapshot`을 갱신합니다.
        SKU 가격이 직전 이력과 다를 때만 `sku_price_history`를 기록합니다.
      operationId: updateSkuSnapshots
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PriceUpdateRequest'
            examples:
              allProducts:
                summary: 전체 상품 대상
                value: {}
              hotProductsOnly:
                summary: 인기 상품만
                value:
                  collection_source: HOT_PRODUCT_QUERY
                  currencies:
                    - KRW
                    - USD
                  requested_by: admin
              filteredByIds:
                summary: 특정 상품만
                value:
                  product_ids:
                    - product-id-1
                    - product-id-2
                  currencies:
                    - USD
                  requested_by: admin
      responses:
        '202':
          description: SKU snapshot 갱신 배치 시작됨
          headers:
            Cache-Control:
              $ref: '#/components/headers/Cache-Control'
            Pragma:
              $ref: '#/components/headers/Pragma'
            Expires:
              $ref: '#/components/headers/Expires'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PriceUpdateAcceptedResponse'
        '400':
          description: 잘못된 요청
          headers:
            Cache-Control:
              $ref: '#/components/headers/Cache-Control'
            Pragma:
              $ref: '#/components/headers/Pragma'
            Expires:
              $ref: '#/components/headers/Expires'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: 관리자 인증 헤더 누락
          headers:
            Cache-Control:
              $ref: '#/components/headers/Cache-Control'
            Pragma:
              $ref: '#/components/headers/Pragma'
            Expires:
              $ref: '#/components/headers/Expires'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: 프리뷰 실패
          headers:
            Cache-Control:
              $ref: '#/components/headers/Cache-Control'
            Pragma:
              $ref: '#/components/headers/Pragma'
            Expires:
              $ref: '#/components/headers/Expires'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /v1/batch/status:
    get:
      tags:
        - batch
      security:
        - AdminApiKey: []
      summary: 배치 상태 조회
      description: 최근 실행된 배치 작업의 상태를 반환합니다.
      operationId: getBatchStatus
      parameters:
        - in: query
          name: job_type
          required: false
          schema:
            type: string
            enum:
              - PRICE_UPDATE
              - SKU_SNAPSHOT_UPDATE
      responses:
        '200':
          description: 배치 상태
          headers:
            Cache-Control:
              $ref: '#/components/headers/Cache-Control'
            Pragma:
              $ref: '#/components/headers/Pragma'
            Expires:
              $ref: '#/components/headers/Expires'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BatchStatusResponse'
        '400':
          description: 잘못된 작업 유형
          headers:
            Cache-Control:
              $ref: '#/components/headers/Cache-Control'
            Pragma:
              $ref: '#/components/headers/Pragma'
            Expires:
              $ref: '#/components/headers/Expires'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: 관리자 인증 헤더 누락
          headers:
            Cache-Control:
              $ref: '#/components/headers/Cache-Control'
            Pragma:
              $ref: '#/components/headers/Pragma'
            Expires:
              $ref: '#/components/headers/Expires'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  headers:
    Cache-Control:
      description: 캐시를 비활성화하는 표준 헤더
      schema:
        type: string
        example: no-store, no-cache, must-revalidate
    Pragma:
      description: HTTP/1.0 클라이언트 캐시를 비활성화하는 호환성 헤더
      schema:
        type: string
        example: no-cache
    Expires:
      description: 응답이 즉시 만료되었음을 나타내는 헤더
      schema:
        type: string
        example: '0'
  securitySchemes:
    AdminApiKey:
      type: apiKey
      in: header
      name: X-Admin-API-Key
      description: |
        모든 `/v1` 어드민 API 호출 시 필요한 헤더.
        `ADMIN_API_KEYS`가 설정되면 해당 값 중 하나와 일치해야 하며, 비어 있으면 하위호환을 위해 non-empty 값만 확인합니다.
  schemas:
...
    APIResponse:
      type: object
      required:
        - status
      properties:
        status:
          type: string
          example: success
        data:
          description: 성공 응답 데이터
        error:
          $ref: '#/components/schemas/ErrorPayload'
    ErrorPayload:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
        message:
          type: string
    ErrorResponse:
      allOf:
        - $ref: '#/components/schemas/APIResponse'
        - type: object
          properties:
            status:
              type: string
              enum:
                - error
            error:
              $ref: '#/components/schemas/ErrorPayload'
    AcceptedMessageResponse:
      allOf:
        - $ref: '#/components/schemas/APIResponse'
        - type: object
          properties:
            data:
              type: object
              required:
                - message
              properties:
                message:
                  type: string
    PriceUpdateRequest:
      type: object
      properties:
        collection_source:
          type: string
          description: 수집 경로 필터
          example: HOT_PRODUCT_QUERY
        market:
          type: string
          description: 마켓 필터
          enum:
            - ALIEXPRESS
            - COUPANG
        product_ids:
          type: array
          description: 특정 내부 product.id 목록. external_product_id가 아니다.
          items:
            type: string
        currencies:
          type: array
          description: 조회/적재 대상 통화. 미지정 시 기존 정책(상품 대표통화 또는 hot product의 KRW+USD)을 사용
          items:
            type: string
            enum:
              - KRW
              - USD
        collected_before:
          type: string
          description: 이 시각 이전에 수집된 상품만 대상
          format: date-time
        force:
          type: boolean
          description: 강제 갱신 여부
        requested_by:
          type: string
          description: 요청자
    HotProductLoadRequest:
      type: object
      properties:
        category_ids:
          type: array
          description: 카테고리 ID 목록
          items:
            type: string
        keywords:
          type: string
          description: 검색 키워드
        sort:
          type: string
          description: 정렬 기준
        min_sale_price:
          type: string
          description: 최소 가격
        max_sale_price:
          type: string
          description: 최대 가격
...
    PriceUpdateAcceptedResponse:
      allOf:
        - $ref: '#/components/schemas/APIResponse'
        - type: object
          properties:
            data:
              type: object
              required:
                - message
                - job
              properties:
                message:
                  type: string
                job:
                  $ref: '#/components/schemas/BatchJobStatus'
...
    BatchStatusResponse:
      allOf:
        - $ref: '#/components/schemas/APIResponse'
        - type: object
          properties:
            data:
              oneOf:
                - $ref: '#/components/schemas/BatchJobStatus'
                - type: object
                  properties:
                    job_type:
                      type: string
                    status:
                      type:
                        - string
                        - 'null'
    BatchJobStatus:
      type: object
      required:
        - job_type
        - status
        - trigger_type
        - filter
        - requested_at
        - total_count
        - success_count
        - fail_count
        - skipped_count
      properties:
        job_type:
          type: string
          enum:
            - PRICE_UPDATE
        status:
          type: string
          enum:
            - QUEUED
            - RUNNING
            - COMPLETED
            - FAILED
        trigger_type:
          type: string
          enum:
            - MANUAL
            - SCHEDULED
        filter:
          $ref: '#/components/schemas/PriceUpdateFilter'
        requested_at:
          type: string
          format: date-time
        started_at:
          type:
            - string
            - 'null'
          format: date-time
        finished_at:
          type:
            - string
            - 'null'
          format: date-time
        total_count:
          type: integer
        success_count:
          type: integer
        fail_count:
          type: integer
        skipped_count:
          type: integer
        last_error:
          type: string
    PriceUpdateFilter:
      type: object
      properties:
        collection_source:
          type: string
        market:
          type: string
          enum:
            - ALIEXPRESS
            - COUPANG
        product_ids:
          type: array
          items:
            type: string
        currencies:
          type: array
          items:
            type: string
            enum:
              - KRW
              - USD
        collected_before:
          type:
            - string
            - 'null'
          format: date-time
        force:
          type: boolean
...